+++ packages/ingest/test/test-granule.js
@@ -587,6 +587,16 @@ test('renameS3FileWithTimestamp renames file', async (t) => {
 
 class TestS3Granule extends s3Mixin(baseProtocol(Granule)) {}
 
+const collectionConfig = {
+  dataType: 'testDataType',
+  version: 'testVersion',
+  files: [
+    {
+      regex: '^[A-Z]|[a-z]+\.txt'
+    }
+  ]
+};
+
 test('ingestFile keeps both new and old data when duplicateHandling is version', async (t) => {
   const sourceBucket = t.context.internalBucket;
   const destBucket = t.context.destBucket;
@@ -597,13 +607,7 @@ test('ingestFile keeps both new and old data when duplicateHandling is version',
   const key = path.join(file.path, file.name);
   const params = { Bucket: sourceBucket, Key: key, Body: randomString() };
   await s3PutObject(params);
-  const collectionConfig = {
-    files: [
-      {
-        regex: '^[A-Z]|[a-z]+\.txt'
-      }
-    ]
-  };
+
   const duplicateHandling = 'version';
   // leading '/' should be trimmed
   const fileStagingDir = '/file-staging';
@@ -643,13 +647,6 @@ test('ingestFile throws error when configured to handle duplicates with error', 
   const params = { Bucket: sourceBucket, Key, Body: 'test' };
   await s3PutObject(params);
 
-  const collectionConfig = {
-    files: [
-      {
-        regex: '^[A-Z]|[a-z]+\.txt'
-      }
-    ]
-  };
   const duplicateHandling = 'error';
   const fileStagingDir = 'file-staging';
   const testGranule = new TestS3Granule(
@@ -668,7 +665,7 @@ test('ingestFile throws error when configured to handle duplicates with error', 
   // first attempt to ingest the file.
   await testGranule.ingestFile(file, destBucket, duplicateHandling);
   const error = await t.throws(testGranule.ingestFile(file, destBucket, duplicateHandling));
-  const destFileKey = path.join(fileStagingDir, file.name);
+  const destFileKey = path.join(fileStagingDir, testGranule.collectionId, file.name);
   t.true(error instanceof errors.DuplicateFile);
   t.is(error.message, `${destFileKey} already exists in ${destBucket} bucket`);
 });